fix: heatmap SQL type mismatch — dow/hour kao i32 umjesto i16
EXTRACT()::smallint - 1 u PostgreSQL-u daje integer (int4) jer 1
literal je int4. sqlx nije mogao decodirati int4 u i16 (smallint),
što je uzrokovalo HTTP 500 grešku na /alarms/heatmap endpointu.

- AlarmHeatmapHour: hour/dow i16 → i32
- SQL: ::smallint → ::integer za hour i dow kolone
- GROUP BY prepiše na ekspresije (ne pozicije) za jednoznačnost
- Dodano eksplicitno ::float8 na AVG

+++ src/models/domain.rs
@@ -719,8 +719,8 @@ pub struct AlarmHeatmapDay {
 /// Prosječna učestalost alarma po satu i danu u tjednu — za hour-of-day heatmap
 #[derive(Debug, Serialize, sqlx::FromRow)]
 pub struct AlarmHeatmapHour {
-    pub hour:  i16,   // 0–23
-    pub dow:   i16,   // 0=Mon, 6=Sun (ISO tjedan − 1)
+    pub hour:  i32,   // 0–23
+    pub dow:   i32,   // 0=Mon, 6=Sun (ISO tjedan − 1)
     pub count: f64,   // prosjek (0.0–1.0) — udio perioda s aktivnim alarmom
 }
